+++ tests/net/ieee802154/l2/src/ieee802154_shell_test.c
@@ -20,7 +20,7 @@ LOG_MODULE_REGISTER(net_ieee802154_mgmt_test, LOG_LEVEL_DBG);
 #include <zephyr/net/net_pkt.h>
 #include <zephyr/shell/shell.h>
 
-#include <ieee802154_frame.h>
+#include <zephyr/net/ieee802154_frame.h>
 
 extern struct net_pkt *current_pkt;
 extern struct k_sem driver_lock;
